docs(scalars): add stories for the text input
