+++ packages/design-system/src/scalars/components/fragments/index.ts
@@ -5,5 +5,4 @@ export * from "./form-label/index.js";
 export * from "./form-message/index.js";
 export * from "./radio-group-field/index.js";
 export * from "./select-field/index.js";
-export * from "./text-field/index.js";
 export * from "./textarea-field/index.js";
